fixes reg e-mail fixes gdpr check
